update xax to multi-model version
